Avoid overriding existing backface-visibility property
If a selector already has a backface-visibility property, this plugin
would add its own anyway, which could override the existing
backface-visibilty property. This commit aims at preventing this from
happening.

Since there is already a backface-visibility property in this case, we
don't actually need to add one of our own.